첨부파일 기능 업로드 삭제기능은 조금 더 손봄

+++ src/main/java/egovframework/cccdms/common/web/FileDownloadController.java
@@ -127,18 +127,16 @@ public class FileDownloadController {
 
 		String atchFileId = (String) commandMap.get("atchFileId");
 		String fileSn = (String) commandMap.get("fileSn");
-
 		Boolean isAuthenticated = CccdmsUserDetailsHelper.isAuthenticated();
-
+		
 		if (isAuthenticated) {
-
 			FileVO fileVO = new FileVO();
 			fileVO.setAtchFileId(atchFileId);
 			fileVO.setFileSn(fileSn);
 			FileVO fvo = fileService.selectFileInf(fileVO);
-
 			File uFile = new File(fvo.getFileStreCours(), fvo.getStreFileNm());
-			long fSize = uFile.length();
+
+			int fSize = (int)uFile.length();
 
 			if (fSize > 0) {
 				String mimetype = "application/x-msdownload";
